update SqrtPrice precision and Percentage type

+++ src/math/types/percentage.rs
@@ -4,4 +4,4 @@ use decimal::*;
 #[derive(Default, Debug, Clone, Copy, PartialEq, Eq, PartialOrd)]
 #[ink::scale_derive(Encode, Decode, TypeInfo)]
 #[cfg_attr(feature = "std", derive(ink::storage::traits::StorageLayout))]
-pub struct Percentage(pub u128);
+pub struct Percentage(pub u64);
